EDMT-73 refactor : useLogin 이름 변경에 따른 수정사항
